Fixed problem under x86_64.
git-svn-id: svn://svn.savannah.gnu.org/myserver/trunk@2958 c4c922ea-a70d-0410-937d-ec1daf91936c

+++ myserver/src/base/process/fork_server.cpp
@@ -93,7 +93,7 @@ int ForkServer::readInt (Socket *sock, int *dest)
  */
 int ForkServer::readString (Socket *sock, char **out)
 {
-  u_long len;
+  int len;
   u_long nbr;
   
   if (sock->read ((char*)&len, 4, &nbr) || nbr < 4)
